+++ katana/sdk/component.py
@@ -55,6 +55,8 @@ class Component(object):
     def set_resource(self, name, callable):
         """Store a resource.
 
+        Callback receives the `Component` instance as first argument.
+
         :param name: Name of the resource.
         :type name: str
         :param callable: A callable that returns the resource value.
@@ -64,7 +66,7 @@ class Component(object):
 
         """
 
-        value = callable()
+        value = callable(self)
         if value is None:
             err = 'Invalid result value for resource "{}"'.format(name)
             raise ComponentError(err)