+++ katana/service.py
@@ -78,15 +78,17 @@ class ServiceServer(ComponentServer):
                 # another service is made.
                 files = get_path(transport, 'files', None)
                 if files:
+                    address = get_path(transport, 'meta/gateway')[1]  # Public gateway address
                     for call in calls:
-                        files_path = '{}/{}/{}'.format(
+                        files_path = '{} {} {} {}'.format(
+                            address,
                             nomap(get_path(call, 'name')),
                             get_path(call, 'version'),
                             nomap(get_path(call, 'action')),
                             )
 
                         # Add flag and exit when at least one call has files
-                        if path_exists(files, files_path):
+                        if path_exists(files, files_path, delimiter=' '):
                             meta += FILES
                             break