+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.1.1] - Unreleased
+### Changed
+- Callback for `Component.set_resource()` now receives the component as
+  the first argument.
 
 ## [1.1.0] - 2017-06-01
 ### Added